final move of the base builders
Both things kinda assume stuff about models and use the FBC interface, so let's
put them to frontend. For FBC one it is actually nice because that does the
flux balance thing right. For the "MMDF" one I just renamed the thing to
concentrations to avoid pretending that it's MMDF-specific (it's not).